sam+heuristic_segmentation/requirements.txt
torch
torchvision
opencv-python
matplotlib
Pillow
segment-anything
scikit-learn
requests
numpy